changing projection to true mercator
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/trunk@91 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/Terrain.h
@@ -58,7 +58,7 @@ public:
 
     bool  isSelectingZone()      {return isSelectionZoneEnCours;}
 
-    bool  getSelectedRectangle (float *x0, float *y0, float *x1, float *y1);
+    bool  getSelectedRectangle (double *x0, double *y0, double *x1, double *y1);
     bool  getSelectedLine      (float *x0, float *y0, float *x1, float *y1);
 
     void setBoatList(QList<boatAccount*> & boat_list) {this->boat_list=&boat_list; update();}
